+++ src/main/java/com/example/demo/dto/deal/DealFilterRequest.java
@@ -1,12 +1,14 @@
 package com.example.demo.dto.deal;
 import com.example.demo.entity.DealStatus;
 
+import java.math.BigDecimal;
+
 
 public record DealFilterRequest(
     String title,
     DealStatus status,
-    Double minAmount,
-    Double maxAmount,
+    BigDecimal minAmount,
+    BigDecimal maxAmount,
     String customerName
 )
 {
